+++ push.py
@@ -2,8 +2,8 @@ import os
 
 def push():
 
-    checkCommand = "pod spec lint SpeedySwift.podspec --verbose --allow-warnings"
-    pushCommand = "pod trunk push SpeedySwift.podspec --verbose --allow-warnings --allow-warnings"
+    checkCommand = "pod spec lint SpeedySwift.podspec --allow-warnings"
+    pushCommand = "pod trunk push SpeedySwift.podspec --allow-warnings --allow-warnings"
     checkRet = os.system(checkCommand)
     if checkRet != 0:
         print("校验出错，请检查spec文件是否配置正确")
